# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artificialic Budget Platform API",
    "description": "API REST provista por nosotros. Consumida por Artificialic (ingest) y dashboard interno.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/analytics/stats": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "M\u00e9tricas y KPIs del dashboard",
        "operationId": "get_stats_v1_analytics_stats_get",
        "parameters": [
          {
            "name": "date_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(today|7d|30d|year|custom)$",
              "default": "30d",
              "title": "Date Range"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/api-keys": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Keys",
        "operationId": "list_keys_v1_admin_api_keys_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Key",
        "operationId": "create_key_v1_admin_api_keys_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/api-keys/{key_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Revoke Key",
        "operationId": "revoke_key_v1_admin_api_keys__key_id__delete",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/api-keys/{key_id}/rotate": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Rotate Key",
        "operationId": "rotate_key_v1_admin_api_keys__key_id__rotate_post",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalyticsStats": {
        "properties": {
          "kpis": {
            "$ref": "#/components/schemas/KpiCards"
          },
          "by_status": {
            "items": {
              "$ref": "#/components/schemas/BreakdownEntry"
            },
            "type": "array",
            "title": "By Status"
          },
          "by_source": {
            "items": {
              "$ref": "#/components/schemas/BreakdownEntry"
            },
            "type": "array",
            "title": "By Source"
          },
          "by_priority": {
            "items": {
              "$ref": "#/components/schemas/BreakdownEntry"
            },
            "type": "array",
            "title": "By Priority"
          },
          "by_assigned": {
            "items": {
              "$ref": "#/components/schemas/BreakdownEntry"
            },
            "type": "array",
            "title": "By Assigned"
          },
          "trend": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Trend"
          },
          "top_services": {
            "items": {
              "$ref": "#/components/schemas/TopServiceEntry"
            },
            "type": "array",
            "title": "Top Services"
          },
          "range": {
            "type": "string",
            "title": "Range"
          }
        },
        "type": "object",
        "required": [
          "kpis",
          "by_status",
          "by_source",
          "by_priority",
          "by_assigned",
          "trend",
          "top_services",
          "range"
        ],
        "title": "AnalyticsStats"
      },
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "examples": [
              "Artificialic Production"
            ]
          },
          "scopes": {
            "items": {
              "type": "string",
              "enum": [
                "budgets:write",
                "budgets:read",
                "budgets:update",
                "admin"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Scopes",
            "examples": [
              [
                "budgets:write"
              ]
            ]
          },
          "rate_limit_per_minute": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Rate Limit Per Minute",
            "default": 60
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "scopes"
        ],
        "title": "ApiKeyCreate"
      },
      "ApiKeyCreated": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "rate_limit_per_minute": {
            "type": "integer",
            "title": "Rate Limit Per Minute"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Plaintext. NO se puede recuperar despu\u00e9s. Gu\u00e1rdelo ahora."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "rate_limit_per_minute",
          "status",
          "created_at",
          "created_by",
          "key"
        ],
        "title": "ApiKeyCreated",
        "description": "Respuesta del POST de creaci\u00f3n. El campo `key` solo aparece aqu\u00ed (y en rotate)."
      },
      "ApiKeyList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ApiKeyOut"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiKeyList"
      },
      "ApiKeyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "rate_limit_per_minute": {
            "type": "integer",
            "title": "Rate Limit Per Minute"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "prefix",
          "scopes",
          "rate_limit_per_minute",
          "status",
          "created_at",
          "created_by"
        ],
        "title": "ApiKeyOut"
      },
      "BreakdownEntry": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "key",
          "count"
        ],
        "title": "BreakdownEntry"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KpiCards": {
        "properties": {
          "total_budgets": {
            "type": "integer",
            "title": "Total Budgets"
          },
          "conversion_rate": {
            "type": "number",
            "title": "Conversion Rate"
          },
          "avg_response_time_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Response Time Hours"
          },
          "potential_revenue": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Potential Revenue"
          }
        },
        "type": "object",
        "required": [
          "total_budgets",
          "conversion_rate",
          "avg_response_time_hours",
          "potential_revenue"
        ],
        "title": "KpiCards"
      },
...
      "TopServiceEntry": {
        "properties": {
          "service_type": {
            "type": "string",
            "title": "Service Type"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          }
        },
        "type": "object",
        "required": [
          "service_type",
          "count",
          "total_amount"
        ],
        "title": "TopServiceEntry"
      },
      "TrendPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "received": {
            "type": "integer",
            "title": "Received"
          },
          "approved": {
            "type": "integer",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "date",
          "received",
          "approved"
        ],
        "title": "TrendPoint"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}